src/third_party/blink/web_tests/external/wpt/web-nfc/idlharness.https.window-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
FAIL NFCWriter interface: existence and properties of interface object assert_own_property: self does not have own property "NFCWriter" expected property "NFCWriter" missing
FAIL NFCWriter interface object length assert_own_property: self does not have own property "NFCWriter" expected property "NFCWriter" missing
FAIL NFCWriter interface object name assert_own_property: self does not have own property "NFCWriter" expected property "NFCWriter" missing
FAIL NFCWriter interface: existence and properties of interface prototype object assert_own_property: self does not have own property "NFCWriter" expected property "NFCWriter" missing
FAIL NFCWriter interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "NFCWriter" expected property "NFCWriter" missing
FAIL NFCWriter interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "NFCWriter" expected property "NFCWriter" missing
FAIL NFCWriter interface: operation push(NDEFMessageSource, NFCPushOptions) assert_own_property: self does not have own property "NFCWriter" expected property "NFCWriter" missing
FAIL NFCWriter must be primary interface of new NFCWriter(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCWriter is not defined"
FAIL Stringification of new NFCWriter(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCWriter is not defined"
FAIL NFCWriter interface: new NFCWriter(); must inherit property "push(NDEFMessageSource, NFCPushOptions)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCWriter is not defined"
FAIL NFCWriter interface: calling push(NDEFMessageSource, NFCPushOptions) on new NFCWriter(); with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCWriter is not defined"
FAIL NFCReader interface: existence and properties of interface object assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface object length assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface object name assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: existence and properties of interface prototype object assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: attribute onreading assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: attribute onerror assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: operation start() assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: operation stop() assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader must be primary interface of new NFCReader(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL Stringification of new NFCReader(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "onreading" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "onerror" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "start()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "stop()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReadingEvent interface: existence and properties of interface object assert_own_property: self does not have own property "NFCReadingEvent" expected property "NFCReadingEvent" missing
FAIL NFCReadingEvent interface object length assert_own_property: self does not have own property "NFCReadingEvent" expected property "NFCReadingEvent" missing
FAIL NFCReadingEvent interface object name assert_own_property: self does not have own property "NFCReadingEvent" expected property "NFCReadingEvent" missing
FAIL NFCReadingEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "NFCReadingEvent" expected property "NFCReadingEvent" missing
FAIL NFCReadingEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "NFCReadingEvent" expected property "NFCReadingEvent" missing
FAIL NFCReadingEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "NFCReadingEvent" expected property "NFCReadingEvent" missing
FAIL NFCReadingEvent interface: attribute message assert_own_property: self does not have own property "NFCReadingEvent" expected property "NFCReadingEvent" missing
FAIL NFCReadingEvent must be primary interface of new NFCReadingEvent("reading", { message: {"url":"/custom/path","records":[{"recordType":"text","data":"Hello World"}]} }) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReadingEvent is not defined"
FAIL Stringification of new NFCReadingEvent("reading", { message: {"url":"/custom/path","records":[{"recordType":"text","data":"Hello World"}]} }) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReadingEvent is not defined"
FAIL NFCReadingEvent interface: new NFCReadingEvent("reading", { message: {"url":"/custom/path","records":[{"recordType":"text","data":"Hello World"}]} }) must inherit property "message" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReadingEvent is not defined"
FAIL NFCErrorEvent interface: existence and properties of interface object assert_own_property: self does not have own property "NFCErrorEvent" expected property "NFCErrorEvent" missing
FAIL NFCErrorEvent interface object length assert_own_property: self does not have own property "NFCErrorEvent" expected property "NFCErrorEvent" missing
FAIL NFCErrorEvent interface object name assert_own_property: self does not have own property "NFCErrorEvent" expected property "NFCErrorEvent" missing
FAIL NFCErrorEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "NFCErrorEvent" expected property "NFCErrorEvent" missing
FAIL NFCErrorEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "NFCErrorEvent" expected property "NFCErrorEvent" missing
FAIL NFCErrorEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "NFCErrorEvent" expected property "NFCErrorEvent" missing
FAIL NFCErrorEvent interface: attribute error assert_own_property: self does not have own property "NFCErrorEvent" expected property "NFCErrorEvent" missing
FAIL NFCErrorEvent must be primary interface of new NFCErrorEvent("error", { error: new DOMException() }); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCErrorEvent is not defined"
FAIL Stringification of new NFCErrorEvent("error", { error: new DOMException() }); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCErrorEvent is not defined"
FAIL NFCErrorEvent interface: new NFCErrorEvent("error", { error: new DOMException() }); must inherit property "error" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCErrorEvent is not defined"
Harness: the test ran to completion.
